Revert "ffi-yajl: update to 2.5.0"
Upstream version has been yanked - without a notification of
course, bravo Chef foundation for this wonderful move.
New gem are labelled -java-universal, not sure if the
bits and pieces of this layer will ever use that one.
But it's okay to stick with 2.4.0 just in case for the
foreseeable future.

This reverts commit c15a87ed53504759e5fc07a8aa8c089bd594b666.