gems: add bindata
BinData is a declarative way to read and write binary file formats.
This means the programmer specifies *what* the format of the binary
data is, and BinData works out *how* to read and write data in this
format. It is an easier (and more readable) alternative to ruby's

See https://rubygems.org/gems/bindata/

Signed-off-by: Carsten Stelling <carsten.stelling@vivavis.com>